info component opens when hovering over heat pump & moved instructions to a separate component
