+++ frontend/src/components/Demo.jsx
@@ -1,16 +1,7 @@
-import { 
-  Box, 
-  Button, 
-  Grid, 
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle, 
-} from '@mui/material';
+import { Box, Button, Grid } from '@mui/material';
 import { useNavigate } from "react-router-dom";
 import { useState } from "react"
-//import EnergyComponent from './EnergyComponent';
+import EnergyComponent from './EnergyComponent';
 import DemoClock from './DemoClock';
 import RealtimeClock from './RealtimeClock';
 import ElectricityPrice from './ElectricityPrice'
@@ -27,6 +18,7 @@ import washingMachineImage from "./../assets/washing_machine.png";
 import solarPanelImage from "./../assets/solar_panel.png";
 import carImage from "./../assets/car.png";
 import electricBoardImage from "./../assets/electric_board.png";
+import Instructions from './Instructions';
 
 const Demo = () => {
   const navigate = useNavigate();
@@ -39,48 +31,14 @@ const Demo = () => {
 
   const [openInstructions, setOpenInstructions] = useState(false);
 
-  const handleCloseInstructions = () => {
-    setOpenInstructions(false);
+  const [componentOpen, setComponentOpen] = useState(false);
+
+  const handleOpen = () => {
+    setComponentOpen(true);
   }
 
-  const instructions = () => {
-    return (
-      <Dialog open={openInstructions} maxWidth={'sm'} fullWidth={true}>
-        
-          <DialogTitle sx={{display: 'flex', justifyContent: 'left', marginX: 2, marginTop: 2}}
-            >{"Liquid AI Demo"}
-          </DialogTitle>
-          <DialogContentText sx={{display: 'flex', justifyContent: 'center', marginX: 5}}>
-            The purpose of this demo is to show how energy consumption can be optimized in a
-            single-family house.
-          </DialogContentText>
-          <DialogTitle sx={{display: 'flex', justifyContent: 'left', marginX: 2, marginTop: 2}}
-            >{"Instructions for use"}
-          </DialogTitle>
-          <DialogContentText sx={{display: 'flex', justifyContent: 'left', marginX: 5}}>
-            You can select what components of the house are included in the demo from the Components
-            dropdown menu. You can inspect individual components by hovering over the component and 
-            clicking on it.<br/>
-            The demo shows the consumption and production of energy for each component and the price
-            of consumed energy by hour. The time range and speed of demo time can be changed from 
-            the dropdown menus.
-          </DialogContentText>
-          <DialogTitle sx={{display: 'flex', justifyContent: 'left', marginX: 2, marginTop: 2}}
-            >{"Source of the data"}
-          </DialogTitle>
-          <DialogContentText sx={{display: 'flex', justifyContent: 'left', marginX: 5}}>
-            Consumption and production data is currently made-up test data for demonstrating purposes.<br/>
-            The price data is fetched from Pörssisähkö API:
-          </DialogContentText>
-          <DialogContent sx={{display: 'flex', justifyContent: 'left', marginX: 2}}>
-            <a href="https://porssisahko.net">Link to the price API (the site is in Finnish)</a>
-          </DialogContent>
-          <DialogActions sx={{display: 'flex', justifyContent: 'center', margin: 2}}>
-            <Button variant='contained' onClick={handleCloseInstructions}>Back to demo</Button>
-          </DialogActions>
-
-      </Dialog>
-    );
+  const handleClose = () => {
+    setComponentOpen(false);
   }
 
   return (
@@ -140,18 +98,33 @@ const Demo = () => {
             }}
           />
           <div>
-          <img
-            src={airHeatPumpImage}
-            alt='heatPump'
-            className='air-heat-pump'
-            style={{
-              position: 'absolute',
-              top: '34%',
-              left: '38%',
-              width: '2%',
-              height: '8%',
-            }}
-          />
+            <img
+              src={airHeatPumpImage}
+              alt='heatPump'
+              className='air-heat-pump'
+              style={{
+                position: 'absolute',
+                top: '34%',
+                left: '38%',
+                width: '2%',
+                height: '8%',
+              }}
+              onMouseEnter={handleOpen}
+              />
+              {componentOpen &&
+              <Box style={{position: 'absolute', zIndex: 1}}>
+                <EnergyComponent 
+                id="2"
+                name="Heat pump"
+                type="consumer"
+                description="Heat pump is used to adjust the temperature inside the house"
+                demoTime={demoTime}
+                handleClose={handleClose}
+              />
+              </Box>
+              }
+                
+          
           <img
             src={freezerImage}
             alt='freezer'
@@ -392,18 +365,17 @@ const Demo = () => {
             <Button 
               variant="contained"
               onClick={() => navigate("/")}>
-            Back
+              Back
             </Button>
             </Grid>
             <Grid item xs={1} style={{minWidth: 200, margin: 5}}>
             <Button 
               variant="contained"
               onClick={() => setOpenInstructions(true)}>
-            More information
+              More information
             </Button>
             </Grid>
-            {instructions()}
-
+            <Instructions openInstructions={openInstructions} setOpenInstructions={setOpenInstructions}/>
           </Grid>
           
         </Grid>
